+++ src/WinFormedge/WebViewCore.HostControl.cs
@@ -47,7 +47,7 @@ partial class WebViewCore
         {
             unsafe
             {
-                DwmSetWindowAttribute((HWND)Container.Handle, DWMWINDOWATTRIBUTE.DWMWA_USE_IMMERSIVE_DARK_MODE, &mode, (uint)sizeof(BOOL)).ThrowOnFailure();
+                DwmSetWindowAttribute((HWND)Container.Handle, DWMWINDOWATTRIBUTE.DWMWA_USE_IMMERSIVE_DARK_MODE, &mode, (uint)sizeof(BOOL));
             }
         }
     }